requirements.txt
Flask[async]==3.0.3
requests==2.32.3
Hypercorn==0.17.3
SQLAlchemy==2.0.35
Hypercorn==0.17.3
beautifulsoup4==4.13.3
lxml==5.3.1
dotenv==0.9.9
tortoise-orm==0.21.6
asyncpg==0.30.0
aerich==0.8.2
tomli_w==1.2.0
tomlkit==0.13.2
quart==0.20.0
google-genai==1.7.0
sqlglot==26.11.1
pandas==2.3.0
pysus==0.15.0
quart-cors==0.8.0